+++ Changes.md
@@ -31,6 +31,7 @@ Fixes
 - Expression : Fixed parsing of Python expressions combining subscripts (`[]`) and `context` methods (#3088, #3613, #5250).
 - ConnectionCreatorWrapper : Fixed bug which forced PlugAdder derived classes to implement `updateDragEndPoint()` unnecessarily.
 - Plug : Fixed bug which caused stale values to be retrieved from the cache for plugs that had been renamed.
+- OpenColorIOTransform : Fixed error processing deep image tiles containing no samples.
 
 API
 ---